fix default creation and change confidence to float

+++ src/dist_s1/data_models/output_models.py
@@ -17,7 +17,7 @@ TIF_LAYER_DTYPES = {
     'GEN-METRIC': 'float32',
     'GEN-DIST-STATUS-ACQ': 'uint8',
     'GEN-METRIC-MAX': 'float32',
-    'GEN-DIST-CONF': 'int16',
+    'GEN-DIST-CONF': 'float32',
     'GEN-DIST-DATE': 'int16',
     'GEN-DIST-COUNT': 'uint8',
     'GEN-DIST-PERC': 'uint8',
